Fix forceWifi on Android 6.0.1
We only need the extreme ACTION_MANAGE_WRITE_SETTINGS permission to forceWifi on Android 6.0.0. On 6.0.1 it was fixed. I have tested that this works on a Samsung device running 6.0.1.
The functionality should remain the same, except we are no longer requesting this additional permission on Android 6.0.1 devices.
Android 6.0.0 devices will still request the permission and Android 7+ devices will still not request the permission.